Move isProbablyUtf8 from kotlin internal to package named internal.
