+++ okhttp-logging-interceptor/src/main/kotlin/okhttp3/logging/HttpLoggingInterceptor.kt
@@ -26,6 +26,7 @@ import okhttp3.OkHttpClient
 import okhttp3.Response
 import okhttp3.internal.http.promisesBody
 import okhttp3.internal.platform.Platform
+import okhttp3.logging.internal.isProbablyUtf8
 import okio.Buffer
 import okio.GzipSource
