+++ okhttp-logging-interceptor/src/test/java/okhttp3/logging/IsProbablyUtf8Test.kt
@@ -15,6 +15,7 @@
  */
 package okhttp3.logging
 
+import okhttp3.logging.internal.isProbablyUtf8
 import okio.Buffer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
